Refactored code a bit to make it more clean
